refactor: 남주님,우혁님 리뷰 반영
