+++ pages/calendar/components/Calendar/DailyCalendar/DailyCalendar.module.scss
@@ -19,7 +19,6 @@
   .all-day-schedules {
     width: 100%;
     height: 3.75rem;
-
     display: flex;
     align-items: center;
     justify-content: center;
@@ -58,7 +57,6 @@
       .label {
         position: absolute;
         bottom: -0.625rem;
-
         width: 3.125rem;
         text-align: center;
         @include caption2-medium($gray700);
