+++ pages/calendar/components/Calendar/WeeklyCalendar/WeeklyCalendar.module.scss
@@ -40,7 +40,7 @@
           align-items: center;
           border-radius: 62.5rem;
           background: $blue500;
-          color: white;
+          color: $white;
         }
       }
     }
